refactor[#226] : settings use...Modal 관련 파일명 변경
